refactor: dedupe test_new_all_reduce helpers into shared module (-587/+30 lines across 2 copies)
Move run_all_reduce_impl and the CRS constants from box/all_post_commit and
box/nightly into blackhole_CI/_all_reduce_helpers.py. The only drift between
the copies (program-cache counting: device-attribute counter with per-run_op
sections vs local counter wrapping the dispatch) is preserved via a
cache_counter_mode parameter bound per file with functools.partial. The
sibling test_all_reduce_2d_fabric{,_turning}.py imports of run_all_reduce_impl
keep resolving. Tails unchanged.
